Move all of the code into js file because typescript sucks
